Refactored implementation by PR comments
